spi: nor-ids: Remove SPI_NOR_HAS_SST26LOCK for sst26wf016
The nor flash sst26wf016 has SPI_NOR_HAS_SST26LOCK which doesn't allow
to write to nor if the nor is not unlocked. This can be done using the
sf unlock command. But because we have disabled this command also on
other nor devices, disable for now also for sst26wf016.