lan969x: Fix SPI access when booting from EMMC
When booting from eMMC then u-boot was not able to access NOR flash.
The reason why was that there was no clock assigned before the probing
the driver. And because of this the probe function failed.
The fix consists in assigning a clock speed before probing the driver
using assigned-clocks property.
Also disable the at91 clocks as they are not needed and also disable
CLK_CCF to be able to use assigned-clocks.