lan969x: Enable CONFIG_DEBUG_UART_BOARD_INIT
The problem was if the boot strap was eMMC/SPI/SDCard without any FC0
traces, then the u-boot will have be missing some prints. The reason is
because much later the pinctrl driver was probed to set the GPIOs. On
top of this we have enabled debug uart, which should print messages even
before setting serial port, but also was not working.
The fix consists in enabling the GPIO inside the board_debug_uart_init.