net: sparx5: Fix usage of mdiobus
We were trying to reuse the mdio bus controller for the PHY ports, so in
case 2 PHYs are on the same MDIO bus, then the 2 PHYs will have the same
bus controller. But the problem was when the second PHY detected that
already exists an MDIO bus controller it didn't use it, it would just
ignore it. Then the port configuration was wrong, because there was a
PHY and a controller but the PHY would not be ever configured.