net: lan969x: Enable QSGMII ports
Get network support also for the network ports connected to the QSGMII
ports. The fix consists in fixing the calendar settings. The settings
are copy from API but there are still issue with ports 9, 13, 17 and 21,
all the other ports seem to be working fine.
The ports are not enabled yet in DT because then the framework will wait
for each of the port to get a link.