configs: lan969x: Enable SPI_FLASH_WINBOND
The following issue have been seen, when booting from NOR flash, u-boot
managed to use quad mode on RX, but when booting from eMMC than it
failed to read. In both cases it worked fine to write to the NOR flash.
The issue seems to be that when booting the u-boot was relying on the
TFA to setup the device in quad mode. So when when booting from eMMC the
TFA was not setting it, while booting from NOR it was setting it.
The fix consists in enabling the quad mode in u-boot. And to enable quad
mode, we can simply enable SPI_FLASH_WINBOND.