lan969x: remove on_mmc_cur
On lan969x there was a callback function for every time when the mmc_cur
env variable was change. And inside this it was changing the mmc_bak.
First problem was that it was selecting the wrong partions, second is
that we have already support for updating mmc_cur and mmc_bak in the
environment.
Therefore remove this function.