configs: lan969x: Remove OF_BOARD_SETUP config
If this is enabled, this will overwrite the meba node in the device
tree. This means that if API will set the pcb to something this will be
overwritten by the pcb env in u-boot, which is not necessary the same.
This option is not enabled on lan966x, so disable it also here.