lan966x: Fix setting boot_source env
With the introduce of tfa_get_boot_source, u-boot is asking the TFA
which is the boot source. The problem here is that if the u-boot runs on
an old TFA which doesn't implement this, it would return UNKNOWN.
There were some fixes for this that in case boot source is UNKNOWN to
get the source by looking at the strapping pins. And this fix was
applied in different places but not for setting the boot_source.
Therefore apply the same fix also for boot_source