Revert "lan969x: Slight change to SPI-NOR env offset with new TFA GPT layout"
This reverts commit 35f45e17247a4f1fa36c1b60af1c43221a28af37.

TFA is building a GPT for NOR that is used by the lmstax. U-boot doesn't
know anything about GPT in NOR and we will not try to synchronize the
NOR format to matches both the u-boot and lmstax requirements.
Therefore remove this and set it as before.

Also another issue is that u-boot uses env and env.bak and with this
change this is not working anymore because the NOR is too small.