lan969x: env: Update env fdt_update
In case we want to apply an overlay it might increase the size of the
fdt and this might required that the fdt to be in a location where it
can be extended. For example if it is in inside the itb then it can't be
expended. Therefore for lan969x move the base fdt to a different place
so we can apply overlays.