spi: Fix unlocking sst26
The sst26 devices can be unlock dynamically. But there is an option
config option CONFIG_SPI_FLASH_UNLOCK_ALL which should unlock the
devices when they are initialized. But this didn't happen for 2 sst26
devices: "sst26vf016b" and "sst26wf016b".
Fix this by setting the correct flags for this devices and make sure
that they are unlocked.