configs: mchp_lan969x_sram_defconfig: Remove FIT and OVERLAY support
Remove these configs from mchp_lan969x_sram_defconfig as this
configuration will not be used to boot a linux kernel and will not need
the overlay support. In this way we reduce the size of the image also.
Otherwise it was noticed that starting this image was failing because it
couldn't allocate enough space.