configs: lan969x: Enable USB and SYSRESET config option
Enable the CONFIG_USB and the other required USB config options, like
USB_DWC3, USB_STORAGE
Enable the CMD_FAT, CMD_USB and CMD_EXT4_WRITE such that we can interact
with the usb.
Enable SYSRESET to be able to reset the board using the GPIO.

While at this run savedefconfig and reorder the config options.