net: mscc_switch: lan969x: Add support for QSGMII interface
Extend the lan969x driver to support also the QSGMII ports. Now it is
possible by using the device tree to choose which port to be used as a
management port. By default it is the RGMII port to keep the backwards
compatibility. The following snippets enables the management port to be
on the port 0 which is used by the QSGMII interface:
port0: port@0 {
    reg = <0>;
    phy-handle = <&phy4>;
    phys = <IF_QSGMII FA_SERDES_TYPE_10G 0>;
};