lan966x: Implement board_quiesce_devices
This function will be called right before jumping to linux kernel. In
this way it is possible to remove/disable whatever it is needed on the
board to have it in a clear state.
In our case we remove the network driver such that it would do a reset
to the switch core.