net: sparx5/lan969x: Do the initialization and reset only once
Now when the driver is probed it would initialize the switch and then
only when the driver is removed it would reset the switch. The reason of
this change is because at every swich reset the GPIO get also reseted
meaning that they can have the correct PU/PD or drive strength.