net: lan966x: Do the initialization and reset only once
Now initialize the RAM and release the reset to the PHYs only first time
when it network driver is used. And then only when the driver is removed
to reset the switch.
The reason of this change is because at every swich reset the GPIO get
also reseted meaning that they can have the correct PU/PD or drive
strength.